Removed UndoEvtMsg to replace it with the more generalised ModifyWhiteboardAnnotationEvtMsg
