+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core2/FromAkkaAppsMsgSenderActor.scala
@@ -83,8 +83,6 @@ class FromAkkaAppsMsgSenderActor(msgSender: MessageSender)
         msgSender.send("from-akka-apps-frontend-redis-channel", json)
       case ClearWhiteboardEvtMsg.NAME =>
         msgSender.send("from-akka-apps-frontend-redis-channel", json)
-      case UndoWhiteboardEvtMsg.NAME =>
-        msgSender.send("from-akka-apps-frontend-redis-channel", json)
       case ModifyWhiteboardAnnotationEvtMsg.NAME =>
         msgSender.send("from-akka-apps-frontend-redis-channel", json)
