+++ bigbluebutton-html5/imports/api/annotations/server/eventHandlers.js
@@ -1,14 +1,12 @@
 import RedisPubSub from '/imports/startup/server/redis';
 import { processForHTML5ServerOnly } from '/imports/api/common/server/helpers';
 import handleWhiteboardCleared from './handlers/whiteboardCleared';
-import handleWhiteboardUndo from './handlers/whiteboardUndo';
 import handleWhiteboardModify from './handlers/whiteboardModify';
 import handleWhiteboardSend from './handlers/whiteboardSend';
 import handleWhiteboardAnnotations from './handlers/whiteboardAnnotations';
 import handleWhiteboardModification from './handlers/whiteboardAnnotationModification';
 
 RedisPubSub.on('ClearWhiteboardEvtMsg', handleWhiteboardCleared);
-RedisPubSub.on('UndoWhiteboardEvtMsg', handleWhiteboardUndo);
 RedisPubSub.on('ModifyWhiteboardAnnotationEvtMsg', handleWhiteboardModify);
 RedisPubSub.on('SendWhiteboardAnnotationEvtMsg', handleWhiteboardSend);
 RedisPubSub.on('GetWhiteboardAnnotationsRespMsg', processForHTML5ServerOnly(handleWhiteboardAnnotations));
